requirements.txt
pandas>=1.3.0
plotly>=5.3.0
numpy>=1.21.0
matplotlib>=3.9.1
seaborn>=0.13.2
scikit-learn>=1.5.1
python-docx>=1.1.2
reportlab>=4.2.5
fpdf>=1.7.2
kaleido>=0.2.1 